profile: rename stats/chartitem.* to stats/statsitem.*
Android uses qmake and apparently that doesn't support source files
with the same name. We already have chart/chartitem.*.
