+++ stats/CMakeLists.txt
@@ -9,8 +9,8 @@ set(SUBSURFACE_STATS_SRCS
 	barseries.cpp
 	boxseries.h
 	boxseries.cpp
-	chartitem.h
-	chartitem.cpp
+	statsitem.h
+	statsitem.cpp
 	chartlistmodel.h
 	chartlistmodel.cpp
 	histogrammarker.h
